+++ OGP1415-JumpingAlien/src/jumpingalien/programs/expressions/GetX.java
@@ -33,7 +33,7 @@ public class GetX extends Expression {
 
 	@Override
 	public Object evaluate(Program program) {
-		return ((GameObject)this.getExpression().evaluate(program)).getHorizontalLocation();
+		return (double)((GameObject)this.getExpression().evaluate(program)).getHorizontalLocation();
 	}
 
 }
